requirements.txt
tqdm==4.62
httpx==0.21
GitPython==3.1
pre-commit==2.17